app/backend/src/utils/seedStockPrice/requirements.txt
requests>=2.28.1
pandas>=1.5.0
psycopg2-binary>=2.9.5
python-dotenv>=1.0.0
types-requests
types-psycopg2